refactor: split account summary statement
