+++ frontend/src/accounts/AccountsPage.tsx
@@ -5,7 +5,7 @@ import { useTaxAdvantagedCategories } from '@/api/taxAdvantagedCategories'
 import CreateAccountModal from '@/components/CreateAccountModal'
 import Filters from '@/accounts/components/filters/Filters'
 import AccountListSection from '@/accounts/components/AccountListSection'
-import AccountSummaryStatement from '@/accounts/components/AccountSummaryStatement'
+import SummaryStatement from '@/accounts/components/summary/SummaryStatement'
 import AccountsMetricsBand from '@/accounts/components/AccountsMetricsBand'
 import ArchivedAccountsSection from '@/accounts/components/ArchivedAccountsSection'
 import TaxAdvantagedLimitsSection from '@/accounts/components/tax-advantaged-limits/TaxAdvantagedLimitsSection'
@@ -53,7 +53,7 @@ export default function AccountsPage() {
       </header>
 
       <div className="space-y-4">
-        <AccountSummaryStatement
+        <SummaryStatement
           error={error}
           isLoading={accountsLoading}
           netWorth={accountSections.netWorth}
