+++ frontend/src/accounts/components/summary/SummaryStatement.tsx
@@ -1,109 +1,11 @@
-import type { ReactNode } from 'react'
-import { AnimatePresence, motion, useReducedMotion } from 'motion/react'
 import { formatCurrency } from '@/utils/formatCurrency'
 import type { FxStatus } from '@/api/shared/fx'
 import { FxStatusTooltip } from '@/components/FxStatusTooltip'
+import { getSummaryFxStatusMessage } from './summaryFxStatus'
+import { SummaryValueFade } from './SummaryValueFade'
+import { SummaryValueSkeleton } from './SummaryValueSkeleton'
 
-function AccountValueSkeleton({
-  label,
-  className,
-}: {
-  label: string
-  className: string
-}) {
-  const shouldReduceMotion = useReducedMotion() ?? false
-
-  return (
-    <div
-      className={`relative overflow-hidden rounded-md ${className}`}
-      role="status"
-      aria-label={label}
-      style={{
-        background: 'var(--app-border)',
-      }}
-    >
-      {!shouldReduceMotion && (
-        <motion.span
-          className="absolute inset-y-0 left-0 w-2/3"
-          aria-hidden
-          initial={{ x: '-140%' }}
-          animate={{ x: '190%' }}
-          transition={{ duration: 1.15, ease: 'easeInOut', repeat: Infinity }}
-          style={{
-            background:
-              'linear-gradient(90deg, transparent, rgba(255, 255, 255, 0.62), transparent)',
-          }}
-        />
-      )}
-    </div>
-  )
-}
-
-function AccountValueFade({
-  loading,
-  skeleton,
-  children,
-}: {
-  loading: boolean
-  skeleton: ReactNode
-  children: ReactNode
-}) {
-  const shouldReduceMotion = useReducedMotion() ?? false
-  const transition = shouldReduceMotion
-    ? { duration: 0 }
-    : { duration: 0.22, ease: [0.42, 0, 0.58, 1] as const }
-
-  return (
-    <AnimatePresence initial={false} mode="popLayout">
-      {loading ? (
-        <motion.div
-          key="skeleton"
-          initial={false}
-          animate={{ opacity: 1 }}
-          exit={{ opacity: 0 }}
-          transition={transition}
-        >
-          {skeleton}
-        </motion.div>
-      ) : (
-        <motion.div
-          key="value"
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-          exit={{ opacity: 0 }}
-          transition={transition}
-        >
-          {children}
-        </motion.div>
-      )}
-    </AnimatePresence>
-  )
-}
-
-function getAccountSummaryFxStatusMessage(fxStatus: FxStatus) {
-  switch (fxStatus.state) {
-    case 'none':
-      return 'All account balances were already in your base currency'
-    case 'complete':
-      return 'Foreign currency account balances were converted into your base currency'
-    case 'incomplete':
-      return 'Some foreign currency accounts could not be converted. Account totals are incomplete and only include accounts with available conversion rates'
-    case 'unavailable':
-      return 'Foreign currency accounts could not be converted. Account totals are incomplete and only include base currency accounts'
-  }
-}
-
-export default function AccountSummaryStatement({
-  error,
-  isLoading,
-  netWorth,
-  totalAssets,
-  totalDebts,
-  assetCount,
-  debtCount,
-  displayCurrency,
-  fxStatus,
-}: {
+type SummaryStatementProps = {
   error: unknown
   isLoading: boolean
   netWorth: number
@@ -113,7 +15,22 @@ export default function AccountSummaryStatement({
   debtCount: number
   displayCurrency: string
   fxStatus: FxStatus
-}) {
+}
+
+/**
+ * Renders account totals, loading placeholders, and FX status for the accounts page summary
+ */
+export default function SummaryStatement({
+  error,
+  isLoading,
+  netWorth,
+  totalAssets,
+  totalDebts,
+  assetCount,
+  debtCount,
+  displayCurrency,
+  fxStatus,
+}: SummaryStatementProps) {
   if (error) {
     return (
       <p className="py-2 font-medium" style={{ color: 'var(--app-negative)' }}>
@@ -139,13 +56,13 @@ export default function AccountSummaryStatement({
             <FxStatusTooltip
               label="Net worth FX status"
               fxStatus={fxStatus}
-              getMessage={getAccountSummaryFxStatusMessage}
+              getMessage={getSummaryFxStatusMessage}
             />
           </div>
-          <AccountValueFade
+          <SummaryValueFade
             loading={isLoading}
             skeleton={(
-              <AccountValueSkeleton
+              <SummaryValueSkeleton
                 label="Loading net worth value"
                 className="h-[3.125rem] w-[min(18rem,76vw)] min-[730px]:h-[3.375rem]"
               />
@@ -159,21 +76,21 @@ export default function AccountSummaryStatement({
             >
               {formatCurrency(netWorth, displayCurrency)}
             </p>
-          </AccountValueFade>
+          </SummaryValueFade>
         </div>
 
         <div className="grid grid-cols-2 gap-4 pb-1.5 min-[730px]:flex min-[730px]:gap-8">
           <div className="text-left min-[730px]:text-right">
             <p className="app-label mb-0.5">Assets</p>
-            <AccountValueFade
+            <SummaryValueFade
               loading={isLoading}
               skeleton={(
                 <div className="grid w-max justify-items-start gap-1.5 min-[730px]:ml-auto min-[730px]:justify-items-end">
-                  <AccountValueSkeleton
+                  <SummaryValueSkeleton
                     label="Loading asset value"
                     className="h-6 w-28"
                   />
-                  <AccountValueSkeleton
+                  <SummaryValueSkeleton
                     label="Loading asset account count"
                     className="h-4 w-20"
                   />
@@ -191,19 +108,19 @@ export default function AccountSummaryStatement({
                   {assetCount} account{assetCount !== 1 ? 's' : ''}
                 </p>
               </>
-            </AccountValueFade>
+            </SummaryValueFade>
           </div>
           <div className="text-right">
             <p className="app-label mb-0.5">Liabilities</p>
-            <AccountValueFade
+            <SummaryValueFade
               loading={isLoading}
               skeleton={(
                 <div className="ml-auto grid w-max justify-items-end gap-1.5">
-                  <AccountValueSkeleton
+                  <SummaryValueSkeleton
                     label="Loading liability value"
                     className="h-6 w-28"
                   />
-                  <AccountValueSkeleton
+                  <SummaryValueSkeleton
                     label="Loading liability account count"
                     className="h-4 w-20"
                   />
@@ -221,7 +138,7 @@ export default function AccountSummaryStatement({
                   {debtCount} account{debtCount !== 1 ? 's' : ''}
                 </p>
               </>
-            </AccountValueFade>
+            </SummaryValueFade>
           </div>
         </div>
       </div>
